requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.77.1
frozenlist==1.3.0
future==0.18.2
h11==0.13.0
idna==3.3
instagram-scraper==1.10.6
joblib==1.1.0
line-bot-sdk==2.2.1
multidict==6.0.2
nltk==3.7
pydantic==1.9.0
regex==2022.4.24
requests==2.27.1
sniffio==1.2.0
starlette==0.19.1
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
yarl==1.7.2
